refactor: prefix reader strategies and writer strategies with Read or Write accordingly.
This eases importing reader and writer strategies of the same kind (source/destination type).
