+++ src/main/java/edu/kit/datamanager/ro_crate/reader/FolderReader.java
@@ -6,7 +6,7 @@ package edu.kit.datamanager.ro_crate.reader;
  * @author Nikola Tzotchev on 9.2.2022 г.
  * @version 1
  *
- * @deprecated Use {@link FolderStrategy} instead.
+ * @deprecated Use {@link ReadFolderStrategy} instead.
  */
 @Deprecated(since = "2.1.0", forRemoval = true)
-public class FolderReader extends FolderStrategy {}
+public class FolderReader extends ReadFolderStrategy {}
