+++ src/main/java/edu/kit/datamanager/ro_crate/writer/ZipWriter.java
@@ -4,7 +4,7 @@ package edu.kit.datamanager.ro_crate.writer;
  * Implementation of the writing strategy to provide a way of writing crates to
  * a zip archive.
  *
- * @deprecated Use {@link ZipStrategy} instead.
+ * @deprecated Use {@link WriteZipStrategy} instead.
  */
 @Deprecated(since = "2.1.0", forRemoval = true)
-public class ZipWriter extends ZipStrategy {}
+public class ZipWriter extends WriteZipStrategy {}
