+++ src/test/java/edu/kit/datamanager/ro_crate/reader/FolderReaderTest.java
@@ -17,7 +17,7 @@ import static org.junit.jupiter.api.Assertions.*;
  * @author Nikola Tzotchev on 9.2.2022 г.
  * @version 1
  */
-class FolderReaderTest implements CommonReaderTest<String, FolderStrategy>
+class FolderReaderTest implements CommonReaderTest<String, ReadFolderStrategy>
 {
   @Override
   public void saveCrate(Crate crate, Path target) throws IOException {
@@ -31,15 +31,15 @@ class FolderReaderTest implements CommonReaderTest<String, FolderStrategy>
   }
 
   @Override
-  public FolderStrategy newReaderStrategyWithTmp(Path tmpDirectory, boolean useUuidSubfolder) {
+  public ReadFolderStrategy newReaderStrategyWithTmp(Path tmpDirectory, boolean useUuidSubfolder) {
     // This strategy does not support a non-default temporary directory
     // and will always use the default one.
     // It also has no state we could make assertions on.
-    return new FolderStrategy();
+    return new ReadFolderStrategy();
   }
 
   @Override
-  public Crate readCrate(FolderStrategy strategy, Path source) throws IOException {
+  public Crate readCrate(ReadFolderStrategy strategy, Path source) throws IOException {
       return new CrateReader<>(strategy)
           .readCrate(source.toAbsolutePath().toString());
   }
