+++ src/test/java/edu/kit/datamanager/ro_crate/writer/ZipWriterTest.java
@@ -5,10 +5,6 @@ import java.nio.file.Path;
 
 import edu.kit.datamanager.ro_crate.Crate;
 import edu.kit.datamanager.ro_crate.RoCrate;
-import edu.kit.datamanager.ro_crate.reader.CommonReaderTest;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -24,13 +20,13 @@ class ZipWriterTest implements
 
     @Override
     public void saveCrateElnStyle(Crate crate, Path target) throws IOException {
-        new CrateWriter<>(new ZipStrategy().usingElnStyle())
+        new CrateWriter<>(new WriteZipStrategy().usingElnStyle())
                 .save(crate, target.toAbsolutePath().toString());
     }
 
     @Override
     public void saveCrateSubdirectoryStyle(RoCrate crate, Path target) throws IOException {
-        new CrateWriter<>(new ZipStrategy().withRootSubdirectory())
+        new CrateWriter<>(new WriteZipStrategy().withRootSubdirectory())
                 .save(crate, target.toString());
     }
 }
